add sidebar for mobile
