+++ apps/dialtone-documentation/docs/.vuepress/theme/components/Sidebar.vue
@@ -19,23 +19,9 @@
 
 <script setup>
 import { useThemeLocaleData } from '@vuepress/plugin-theme-data/client';
-import SidebarItem from '../components/SidebarItem.vue';
-import { computed } from 'vue';
-import { useRoute } from 'vue-router';
-import { usePageData } from '@vuepress/client';
+import SidebarItem from './SidebarItem.vue';
+import { useSidebarItems } from '../composables/useSidebarItems';
 
-const route = useRoute();
 const items = useThemeLocaleData().value.sidebar;
-const sidebarItems = computed(() => {
-  const key = Object.keys(items).filter(item => route.path.includes(item.replace(/\/$/, '')));
-  if (!items[key]) return [];
-  if (Array.isArray(items[key])) return items[key] || [];
-  return [{
-    isSinglePage: true,
-    text: items[key].text,
-    children: usePageData().value.headers.map(item => {
-      return { text: item.title, link: item.link };
-    }),
-  }];
-});
+const sidebarItems = useSidebarItems(items);
 </script>
